- moved user verification function ("filter") from User => Boolean to User => Future[Boolean], since sometimes you need to find out more about a user to see if he may access the resource and a synchronous call simply is not enough - extracted used implicit ExecutionContext and made it available as an implicit parameter to let the user choose which ExecutionContext should be used
