+++ app/org/zalando/hutmann/authentication/OAuth2Action.scala
@@ -6,13 +6,13 @@ import com.typesafe.config.Config
 import org.zalando.hutmann.logging.{ Context, Logger, RequestContext }
 import play.api.Play.{ configuration, current }
 import play.api.http.Status
-import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import play.api.libs.ws.{ WS, WSRequest }
 import play.api.mvc.Results._
 import play.api.mvc._
 
-import scala.concurrent.Future
+import scala.concurrent.{ ExecutionContext, Future }
 import scala.concurrent.duration._
+import scala.language.implicitConversions
 import scala.util.control.NonFatal
 import scala.util.matching.Regex
 import scala.util.{ Failure, Success, Try }
@@ -24,10 +24,10 @@ import scala.util.{ Failure, Success, Try }
   * @param config The config section containing the OAuth server information
   */
 class OAuth2Action(
-  val filter:         User => Boolean = { user: User => true },
-  val autoReject:     Boolean         = true,
-  val requestTimeout: Duration        = 1.seconds
-)(implicit config: Config)
+  val filter:         User => Future[Boolean] = { user: User => Future.successful(true) },
+  val autoReject:     Boolean                 = true,
+  val requestTimeout: Duration                = 1.seconds
+)(implicit config: Config, ec: ExecutionContext)
     extends ActionBuilder[UserRequest] {
 
   val url = config.getString("org.zalando.hutmann.authentication.oauth2.tokenInfoUrl")
@@ -105,18 +105,19 @@ class OAuth2Action(
   /**
     * Scores the types of problems that can occur and maps them to appropriate user-usable ones.
     */
-  def tokenResultScoring[A](implicit context: Context): (Either[OAuth2Error, User]) => Either[AuthorizationProblem, User] = {
+  def tokenResultScoring[A](implicit context: Context): (Either[OAuth2Error, User]) => Future[Either[AuthorizationProblem, User]] = {
     case Right(user) =>
-      val isValidUser = filter(user)
-      if (isValidUser) {
-        Right(user)
-      } else {
-        Logger.info("User is authorized but doesn't fit the given filter")
-        Left(InsufficientPermissions(user))
+      filter(user).map { isValidUser =>
+        if (isValidUser) {
+          Right(user)
+        } else {
+          Logger.info("User is authorized but doesn't fit the given filter")
+          Left(InsufficientPermissions(user))
+        }
       }
     case Left(failure) =>
       Logger.info(s"Failed to validate token: $failure")
-      Left(NoAuthorization)
+      Future.successful(Left(NoAuthorization))
   }
 
   /**
@@ -137,7 +138,7 @@ class OAuth2Action(
             validateToken(token)
           case other => Future.successful(other)
         }
-        val futureUser = futureToken.map(tokenResultScoring).recoverWith({
+        val futureUser = futureToken.flatMap(tokenResultScoring).recoverWith({
           case _: TimeoutException => Future.successful(Left(AuthorizationTimeout))
           case e                   => throw e
         })
@@ -146,16 +147,6 @@ class OAuth2Action(
         Logger.info("No authorization founder in 'Authorization' header or 'access_token' query parameter. Rejecting access.")
         Future.successful(new UserRequest(Left(NoAuthorization), request))
     }
-  }
-
-  /**
-    * Creates a new instance of `OAuth2` using the predicate to filter the user
-    *
-    * @param predicate a function that inspects the user if he should get access or not
-    * @return
-    */
-  def withUserFilter(predicate: (User) => Boolean): ActionBuilder[UserRequest] = {
-    new OAuth2Action(predicate)(config)
   }
 
   override def invokeBlock[A](request: Request[A], block: (UserRequest[A]) => Future[Result]): Future[Result] = {
@@ -198,12 +189,22 @@ class OAuth2Action(
 }
 
 object OAuth2Action {
+  def withUserFilter(
+    filter: User => Boolean
+  )(implicit ec: ExecutionContext): OAuth2Action =
+    apply({ user: User => Future.successful(filter(user)) })
+
   def apply(
-    filter:         User => Boolean = { user: User => true },
-    autoReject:     Boolean         = true,
-    requestTimeout: Duration        = 1.second
-  ): OAuth2Action =
-    new OAuth2Action(filter, autoReject, requestTimeout)(configuration.underlying)
+    filter: User => Future[Boolean] = { user: User => Future.successful(true) }
+  )(implicit ec: ExecutionContext): OAuth2Action =
+    apply(filter, autoReject = true, 1.second)
+
+  def apply(
+    filter:         User => Future[Boolean],
+    autoReject:     Boolean,
+    requestTimeout: Duration
+  )(implicit ec: ExecutionContext): OAuth2Action =
+    new OAuth2Action(filter, autoReject, requestTimeout)(configuration.underlying, ec)
 }
 
 /** Holds the regular expression patterns that are used for reading the tokens from headers and query parameters. Extracted here due to performance reasons.*/
